MasterClass-MCP-Agent/requirements.txt
mcp
openai
python-dotenv
fastapi
uvicorn[standard]
jinja2
aiofiles
requests 